// src/app/models/pagos_header.ts
export interface PagoHeader {
  pag_com_ref: boolean;
  bas_tot:number;
  id: number;
  prv: number;
  est: string;
  num_ped_prv: string;
  fch: string;
  num_ped: number;
  bas_gen: number;
  id_roi_usr: number;
  id_plan: number;
  vta_fac: number;
  id_com_ped_lin:number;
  selected: boolean;
}